added spinner

+++ drive-server/src/controllers/FileController.ts
@@ -2,13 +2,11 @@ import { Request, Response } from "express";
 import { FileService } from "../services/FileService";
 import multer from 'multer';
 import fs from 'fs';
-import File,{ IFile } from "../models/file-model";
+import File, { IFile } from "../models/file-model";
 
-
-const upload = multer({dest: 'uploads/'})
+const upload = multer({ dest: 'uploads/' })
 
 export class FileController {
-
   private fileService: FileService;
 
   constructor(bucketName: string) {
@@ -16,26 +14,32 @@ export class FileController {
   }
 
   public async uploadFile(req: Request, res: Response) {
-    if(!req.file) {
+    if (!req.file) {
       return res.status(400).send({ error: 'No file provided' });
     }
 
     try {
+      // Check if file with the same original name already exists
+      const existingFile = await File.findOne({ originalname: req.file.originalname });
+      if (existingFile) {
+        return res.status(500).send({ error: 'File with the same name already exists' });
+      }
+
       const fileName = req.file.originalname;
       const signedUrl = await this.fileService.generateUploadSignedUrl(fileName);
       await this.fileService.uploadFileToGCS(signedUrl, req.file.path);
       await this.saveFileMetadata(req);
       fs.unlinkSync(req.file.path);
-      res.status(200).send({message: 'File uploaded successfully'});
-    } catch(error) {
+      res.status(200).send({ message: 'File uploaded successfully' });
+    } catch (error) {
       if (req.file && fs.existsSync(req.file.path)) {
         fs.unlinkSync(req.file.path);
       }
-      res.status(500).json({ error: 'Error in uploading file' });
+      res.status(500).json({ error: error || 'Error in uploading file' });
     }
   }
 
-  private async saveFileMetadata(req: Request) : Promise<void> {
+  private async saveFileMetadata(req: Request): Promise<void> {
     console.log(req.user)
     try {
       const fileData = new File({
@@ -45,9 +49,9 @@ export class FileController {
         uploaderName: req.user.name
       });
       await fileData.save();
-    } catch(error) {
-        console.error('Error saving file metadata:', error);
-        throw error; // Rethrow the error to be caught by the caller
-    } 
+    } catch (error) {
+      console.error('Error saving file metadata:', error);
+      throw error; // Rethrow the error to be caught by the caller
+    }
   }
 }
